Refactor in-app structure: separate classes into individual files, introduce new enums, and update imports accordingly
