+++ src/inApp/enums/IterableInAppCloseSource.ts
@@ -1,0 +1,8 @@
+// TODO: Add description
+export enum IterableInAppCloseSource {
+  back = 0,
+  link = 1,
+  unknown = 100,
+}
+
+export default IterableInAppCloseSource;
