+++ src/inApp/enums/IterableInAppContentType.ts
@@ -1,0 +1,8 @@
+// TODO: Add description
+export enum IterableInAppContentType {
+  html = 0,
+  alert = 1,
+  banner = 2,
+}
+
+export default IterableInAppContentType;
