+++ src/inApp/enums/IterableInAppDeleteSource.ts
@@ -1,0 +1,8 @@
+// TODO: Add description
+export enum IterableInAppDeleteSource {
+  inboxSwipe = 0,
+  deleteButton = 1,
+  unknown = 100,
+}
+
+export default IterableInAppDeleteSource;
