+++ src/inApp/enums/IterableInAppLocation.ts
@@ -1,0 +1,7 @@
+// TODO: Add description
+export enum IterableInAppLocation {
+  inApp = 0,
+  inbox = 1,
+}
+
+export default IterableInAppLocation;
