+++ src/inApp/enums/IterableInAppShowResponse.ts
@@ -1,0 +1,9 @@
+// TODO: Add description
+export enum IterableInAppShowResponse {
+  /** Show in-app */
+  show = 0,
+  /** Skip showing in-app */
+  skip = 1,
+}
+
+export default IterableInAppShowResponse;
