+++ src/inApp/index.ts
@@ -1,3 +1,7 @@
-export * from './IterableInAppClasses';
+export * from './enums';
+export * from './IterableHtmlInAppContent';
 export * from './IterableInAppManager';
 export * from './IterableInAppMessage';
+export * from './IterableInAppTrigger';
+export * from './IterableInboxMetadata';
+export * from './types';
