+++ src/inApp/types/IterableInAppContent.ts
@@ -1,0 +1,6 @@
+import type { IterableInAppContentType } from '../enums';
+
+// TODO: Add description
+export interface IterableInAppContent {
+  type: IterableInAppContentType;
+}
